+++ lib/rble/cli/formatters/text.rb
@@ -8,8 +8,8 @@ module RBLE
       class Text
         def device(device)
           name = truncate(device.name || company_label(device) || "(unknown)", 25)
-          rssi = device.rssi || 0
-          puts format("%-17s  %-25s  %4d dBm  %s", device.address, name, rssi, device.address_type)
+          rssi_str = device.rssi ? format("%4d dBm", device.rssi) : "     N/A"
+          puts format("%-17s  %-25s  %s  %s", device.address, name, rssi_str, device.address_type)
         end
 
         def status(info)